Refactor CloudBackupSupervisor into submodules
Break up the large CloudBackupSupervisor into focused submodules (enable, disable, cloud_only, verification, tests) and reorganize imports. Delegate upload/sync/write responsibilities to dedicated workers and add new supervisor APIs for restore-with-events, cancel restore, upload scheduling/resumption, sync-health requests, pending verification caching, and runtime state clearing. Remove many monolithic verification/disable/enable flows from the supervisor and replace them with calls into the new modules/workers to simplify control flow and make operations more testable.
